Bug 793146 - automatic filename to allow immediate recording
This will enable recording with 1 single click.
Added automatic filename creation on recording based on date and time

This option can be enabled in preferences, together with the direction
that should be recorded. A extension for automatic file name creation
can be configured, too.

When enabled in preferences the record is stored in home directory with the
name scheme moserial_YYYY-mm-dd_HH-MM-SS.txt